DEBAT_AI/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
asyncpg
python-dotenv
openai
pydantic
